requirements.txt
# Core FastAPI and Web Framework
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
starlette==0.41.3

# Data Validation and Processing
pydantic==2.9.2
pydantic_core==2.23.4

# Web Scraping and Browser Automation
selenium==4.34.2
undetected-chromedriver==3.5.5
beautifulsoup4==4.12.3
lxml==5.3.0
requests==2.32.3

# AI and Google Services
google-generativeai
google-adk

# Database and Supabase
supabase==2.10.0
postgrest==0.18.0
gotrue==2.11.0
realtime==2.0.6
storage3==0.9.0
supafunc==0.7.0

# Vietnamese Stock Data
vnstock==3.2.6
vnstock_ezchart==0.0.2

# Environment and Configuration
python-dotenv==1.0.1

# HTTP and Networking
httpx==0.27.2
aiohttp==3.11.11
urllib3==2.5.0

# Data Analysis and Processing
pandas==2.3.0
numpy==1.26.4

# Date and Time Handling
python-dateutil==2.9.0.post0

# Async Support
asyncio
aiosqlite==0.20.0

# Text Processing
regex==2024.9.11

# Development and Testing
pytest
pytest-asyncio